arch(types): Rename BrsType -> BrsNumber
It only handles numbers, and it exists in src/brsTypes/ so the old name
was potentially confusing.
