+++ src/brsTypes/BrsNumber.ts
