+++ src/brsTypes/index.ts
@@ -1,3 +1,3 @@
-export * from "./BrsType";
+export * from "./BrsNumber";
 export * from "./Float";
 export * from "./Int32";
